Added new guassian field generation methods to generate power spectra with fixed min and max length scales,  added a more efficient power spectrum util
